refactor: move const to lib folder
