+++ lib/const/const.js
@@ -3,4 +3,3 @@ export const EXIT = 'Thank you for using File Manager';
 export const WAITING = '🤖 Waiting for your commands...';
 export const INVALID_INPUT = '📛 Invalid input';
 export const ERROR = '💥 Operation failed';
-export const EXIT_COMMAND = '.logExit';
